+++ tool/travis.sh
@@ -4,6 +4,7 @@ set -e
 
 cd "$1"
 
+echo "> entered package $1"
 echo '> pub get ================================================================'
 pub get
 
@@ -12,10 +13,18 @@ if [ -f "package.json" ]; then
     npm install
 fi
 
-echo "> pub run test -r expanded ==============================================="
-pub run test -r expanded
+if [ "$2" = "node" ]; then
+    echo "> pub run build_runner test (dartdevc) ============================="
+    pub run build_runner test --output=build/ --low-resources-mode --verbose -- -r expanded
 
-echo '> dartfmt -n --set-exit-if-changed . ====================================='
+    echo "> pub run build_runner test (dart2js) ============================="
+    pub run build_runner test --define="build_node_compilers|entrypoint=compiler=dart2js" --output=build/ --low-resources-mode --verbose -- -r expanded
+else
+    echo "> pub run test -r expanded ==============================================="
+    pub run test -r expanded
+fi
+
+echo "> dartfmt -n --set-exit-if-changed . ====================================="
 dartfmt -n --set-exit-if-changed .
 
 echo "> dartanalyzer --fatal-infos --fatal-warnings . =========================="